refactor fetching saving and loading history to service architecture
