+++ CosmoPic/Models/DataStoreNew.swift
@@ -9,15 +9,21 @@ import Foundation
 
 class DataStoreNew: ObservableObject {
   @Published var photo: Photo?
+  @Published var history: [Photo] = []
   @Published var isLoading = false
 
   @Published var errorIsPresented = false
   @Published var error: Error?
 
   let photoAPIService: PhotoAPIServiceProtocol
+  let historyAPIService: HistoryAPIServiceProtocol
 
-  init(photoAPIService: PhotoAPIServiceProtocol = PhotoAPIService()) {
+  init(
+    photoAPIService: PhotoAPIServiceProtocol = PhotoAPIService(),
+    historyAPIService: HistoryAPIServiceProtocol = HistoryAPIService()
+  ) {
     self.photoAPIService = photoAPIService
+    self.historyAPIService = historyAPIService
   }
 
   @MainActor
@@ -54,5 +60,40 @@ class DataStoreNew: ObservableObject {
 
   func getLocalFileURL(forFilename filename: String) -> URL {
     return FileManager.documentsDirectoryURL.appendingPathComponent(filename)
+  }
+
+  @MainActor
+  func getHistory() async {
+    do {
+      let dateFormatter = DateFormatter()
+      dateFormatter.dateFormat = "yyyy-MM-dd"
+      let currentDate = Date()
+      guard let oneMonthAgo = Calendar.current.date(byAdding: .month, value: -1, to: currentDate) else { return }
+
+      let todayString = dateFormatter.string(from: currentDate)
+      let startDate = dateFormatter.string(from: oneMonthAgo)
+
+      let historyFilePath = FileManager.documentsDirectoryURL
+        .appendingPathComponent("\(todayString)-history.json")
+
+      if FileManager.default.fileExists(atPath: historyFilePath.path) {
+        history = try historyAPIService.loadHistory(for: todayString)
+      } else {
+        let fetchedHistory = try await historyAPIService.fetchHistory(starting: startDate, ending: todayString)
+
+        let updatedHistory = try await historyAPIService.updatePhotosWithLocalURLs(
+          fetchedHistory,
+          dateFormatter: dateFormatter
+        )
+
+        try historyAPIService.saveHistory(updatedHistory, for: todayString)
+
+        history = updatedHistory
+      }
+    } catch {
+      self.error = error
+      errorIsPresented = true
+    }
+    isLoading = false
   }
 }
